ImportanceSampling

Documentation for ImportanceSampling.

ImportanceSampling.AMIS

ImportanceSampling.AnalyticalFitter

ImportanceSampling.NormalProposal

ImportanceSampling.ProposalDistribution

ImportanceSampling.SamplingMethod

Base.rand

Distributions.logpdf

ImportanceSampling.collect_samples

ImportanceSampling.collect_weights

ImportanceSampling.decode_params

ImportanceSampling.estimate_parameters!

ImportanceSampling.exp_weights

ImportanceSampling.fit_distribution!

ImportanceSampling.initial_params

ImportanceSampling.loglikelihood

ImportanceSampling.optimize_multistart

ImportanceSampling.resample

ImportanceSampling.set_params!

ImportanceSampling.AMIS — Type

AMIS(; kwargs...)

Performs the AMIS (adaptive multiple importance sampling) algorithm. Returns (T * N) importance samples with weights.

Keywords

T::Int64: Number of iterations.

N::Int64: Number of samples in each iteration.

ImportanceSampling.AnalyticalFitter — Type

Estimates the proposal distribution parameters analytically.

Only works with proposal distributions which implement estimate_parameters!.

ImportanceSampling.NormalProposal — Type

Normal proposal distribution parametrized by the mean vector and the covariance matrix.

ImportanceSampling.ProposalDistribution — Type

Abstract type for proposal distributions.

Each proposal distribution has to implement the following API:

initial_params(::ProposalDistribution, count::Int; options::ISOptions = ISOptions())

loglikelihood(::ProposalDistribution, xs::AbstractMatrix{<:Real}; options::ISOptions = ISOptions())

set_params!(::ProposalDistribution, θ::AbstractVector{<:Real}; options::ISOptions = ISOptions())

The ProposalDistribution should be parametrized by a vector of real-valued numbers.

(This means that for example the normal distribution cannot be parametrized by the individual elements of its covariance matrix directly as that would not necesarilly result in a positive-definite covariance matrix. Instead, suitable parameter transformations have to be implemented as a part of the functions of the ProposalDistribution API, so that the DistributionFitter can work with real-valued parameters.)

ImportanceSampling.SamplingMethod — Method

Sample from the distribution given by the logpdf log_π using the given ProposalDistribution and DistributionFitter.

Return the samples (as a column-wise matrix) and their respective weights (as a vector).

The weights are not necesarilly normalized.

Base.rand — Method

Sample count independent samples from the ProposalDistribution with its current parameters.

Distributions.logpdf — Method

Return the log likelihood of the given data point x under the distribution with the current parameters.

ImportanceSampling.collect_samples — Method

Return all samples from the given iterations ts.

ImportanceSampling.collect_weights — Method

Return weights of the samples from the iterations ts.

ImportanceSampling.decode_params — Method

Decode the real-valued parameters θ into valid mean vector and covariance matrix.

ImportanceSampling.estimate_parameters! — Method

Analytically compute the optimal estimate of the distribution parameters according to the given data xs.

This function is a part of the optional API of the ProposalDistribution and may not be implemented for every distribution.

ImportanceSampling.exp_weights — Method

Exponentiate the given weights in a numerically stable way.

ImportanceSampling.fit_distribution! — Method

Find the optimal parameters of the given ProposalDistribution that best fit the given data xs.

ImportanceSampling.initial_params — Method

Return some initial parameter values for the likelihood maximization.

The parameter values may be some reasonable defaults or random values.

ImportanceSampling.loglikelihood — Method

Return a function mapping the distribution parameters θ to the log-pdf of the given data xs.

ImportanceSampling.optimize_multistart — Method

Run the optimization (maximization) procedure contained within the optimize argument multiple times and return the best local optimum found this way.

Return results of all successful runs (not just the best one) if the kwargs return_all is set to true.

Throws an error if all optimization runs fail.

ImportanceSampling.resample — Method

xs = resample(xs::AbstractMatrix{<:Real}, ws::AbstractVector{<:Real}, count::Int)

Resample count samples from the given data set xs weighted by the given weights ws with replacement to obtain a new un-weighted data set.

Some data points may repeat in the resampled data set. Increasing the sample size of the initial data set may help to reduce the number of repetitions.

ImportanceSampling.set_params! — Method

Set the parameters of the given ProposalDistribution to the given values θ.